Start of tests for maas2Instance
